pyproject.toml
[tool.poetry]
name = "sql"
version = "0.1.0"
description = ""
authors = ["Gonzalo Ayuso <gonzalo123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain-openai = "^0.1.15"
langchain-community = "^0.2.7"
click = "^8.1.7"
langchain-groq = "^0.1.6"
dbutils3-gonzalo123 = "^3.0.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
bandit = "^1.7.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
